Move non-component modules out of src/components into src/lib
Four non-tsx files lived under src/components despite being plain logic,
not components. Relocate them to src/lib so the component tree holds only
components:

- use-vto-requests.ts (shared VTO request hook) -> src/lib/
- availability.ts (pure outfit/composition-rule logic) -> src/lib/
  fitting-room-outfit.ts; renamed so the concern is clear once it's out
  of the fitting-room/ directory.
- product-sizing-types.ts (VtoProductData view-model types) and
  fitting-room/product-data.ts (its builders) -> merged into
  src/lib/product.ts, co-located with loadProductData.

Also rename fitting-room.ts -> fitting-room-storage.ts to distinguish the
localStorage persistence layer from fitting-room-data.ts (React data
resolution) and fitting-room-outfit.ts (composition rules) — three files
sharing the fitting-room feature but each a distinct layer.

product.ts now type-imports ResolvedFittingRoomItem from
fitting-room-data.ts, which value-imports product.ts. The product side is
`import type`, fully erased, so there is no runtime import cycle.

Pure move/merge — no behavior change. npm run check + build pass;
single-garment VTO smoke-tested in the browser (overlay mounts, frames
render, no console errors).
