+++ src/components/overlays/fitting-room/desktop-layout.tsx
@@ -3,7 +3,7 @@ import { LinkT } from '@/components/link'
 import { TfrIcon } from '@/lib/asset'
 import { ResolvedFittingRoom, ResolvedFittingRoomItem } from '@/lib/fitting-room-data'
 import { useCss } from '@/lib/theme'
-import { Availability } from './availability'
+import { Availability } from '@/lib/fitting-room-outfit'
 import { AvatarControls } from './avatar-controls'
 import { AvatarPane } from './avatar-pane'
 import { CardRail } from './card-rail'
