+++ src/components/overlays/fitting-room/detail-accordion-item.tsx
@@ -6,12 +6,11 @@ import { ItemFitText } from '@/components/item-fit-text'
 import { LinkT } from '@/components/link'
 import { SizeSelector } from '@/components/size-selector'
 import { Text, TextT } from '@/components/text'
-import { VtoProductData } from '@/components/product-sizing-types'
 import { ResolvedFittingRoomItem } from '@/lib/fitting-room-data'
 import { useTranslation } from '@/lib/locale'
+import { buildVtoProductDataFromResolved, VtoProductData } from '@/lib/product'
 import { useCss } from '@/lib/theme'
 import { Chevron } from './chevron'
-import { buildVtoProductDataFromResolved } from './product-data'
 
 export type DetailMode = 'compact' | 'expanded'
 export type Platform = 'desktop' | 'mobile'
