+++ src/components/widgets/add-to-fitting-room-compact.tsx
@@ -1,5 +1,5 @@
 import { FittingRoomIcon } from '@/lib/asset'
-import { toggleFittingRoomItem } from '@/lib/fitting-room'
+import { toggleFittingRoomItem } from '@/lib/fitting-room-storage'
 import { useTranslation } from '@/lib/locale'
 import { getLogger } from '@/lib/logger'
 import { getStaticData, useMainStore } from '@/lib/store'
