+++ src/index.tsx
@@ -6,7 +6,7 @@ import { _init as initApi } from '@/lib/api'
 import { _init as initAsset } from '@/lib/asset'
 import { getConfig, EnvName } from '@/lib/config'
 import { _init as initFirebase, getAuthManager } from '@/lib/firebase'
-import { _init as initFittingRoom } from '@/lib/fitting-room'
+import { _init as initFittingRoom } from '@/lib/fitting-room-storage'
 import { i18n } from '@/lib/locale'
 import { _init as initLogger, getLogger } from '@/lib/logger'
 import { _init as initProduct } from '@/lib/product'
